rag_ai/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain-chroma==0.2.3
langchain-community==0.3.12
langchain-core==0.3.52
langchain-ollama==0.3.2
langchain-text-splitters==0.3.8
chromadb==0.5.23
sentence-transformers==2.2.2
pydantic>=2.5.0
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
